Move the debug marker initialization part to another file (#2288)
* Move the debug marker initialization part to 'debug.cpp'
