+++ localization/pose_estimator/ndt_scan_matcher/ndt_scan_matcher/CMakeLists.txt
@@ -15,6 +15,7 @@ find_package(ament_cmake_auto REQUIRED)
 ament_auto_find_build_dependencies()
 
 ament_auto_add_executable(ndt_scan_matcher
+  src/debug.cpp
   src/ndt_scan_matcher_node.cpp
   src/ndt_scan_matcher_core.cpp
   src/util_func.cpp
